lto: Fix SegFault in ICF caused by missing body
During LTO symbol merging, weak symbols may be resolved to external
definition.
We reset the symbol, so the body might be released in unreachability
pass. But we didn't mark the symbol with body_removed, so ICF assumed
the body was still there causing SegFault.

gcc/lto/ChangeLog:

	* lto-symtab.cc (lto_symtab_merge_symbols): Set body_removed
	for symbols resolved outside of IR.

gcc/testsuite/ChangeLog:

	* gcc.dg/lto/attr-weakref-2_0.c: New test.
	* gcc.dg/lto/attr-weakref-2_1.c: New test.