c-family: Fix ICE with -MD and -fdeps-format sharing output [PR121864]
When -MD, -fdeps-format=p1689r5 and -save-temps are used without
explicit output files, they default to the same stream, which is
invalid. The error message attempted to print fdeps_file, but this is
NULL in this case, causing an ICE.

Use out_fname as a fallback when fdeps_file is NULL to avoid the ICE
and provide a meaningful error message.

Fix suggested by Andrew Pinski.

	PR c++/121864

gcc/c-family/ChangeLog:

	* c-opts.cc (c_common_finish): Use out_fname as fallback when
	fdeps_file is NULL in error message.

Signed-off-by: Egas Ribeiro <egas.g.ribeiro@gmail.com>
Reviewed-by: Jason Merrill <jason@redhat.com>