ifcvt: Fix noce_try_cond_zero_arith after get_base_reg change [PR123267]
A few fixes are needed after the change to get_base_reg of
r16-6333-gac64ceb33bf05b. First we need to use the correct target mode
of the operand, this means if we are doing a subreg of QI mode, using
QImode for the conditional move.
Second we also need to use the original operands instead of the ones
removing the subreg still.

Pushed as obvious after a bootstrap/test on x86_64-linux-gnu.

	PR rtl-optimization/123267
gcc/ChangeLog:

	* ifcvt.cc (noce_try_cond_zero_arith): Pass the original operands
	of a instead of the stripped off values. The mode of the operand
	which is being used.

gcc/testsuite/ChangeLog:

	* gcc.dg/torture/pr123267-1.c: New test.

Signed-off-by: Andrew Pinski <andrew.pinski@oss.qualcomm.com>