LRA: Fix eliminate regs into a subreg inside a debug insn [PR123295]
So the problem here is during LRA we are eliminating argp and trying to
simplify the RTL as we go but inside a debug insn, almost all subreg
are valid due to gen_lowpart_for_debug done during debug insn simplification.
So simplify_gen_subreg will fail on some subregs and return null.
This causes problems later on. The solution is create a raw SUBREG
like what is done in lra_substitute_pseudo for debug insns.

Bootstrapped and tested on x86_64-linux-gnu.

	PR rtl-optimization/123295
gcc/ChangeLog:

	* lra-eliminations.cc (lra_eliminate_regs_1): For a debug
	insn, create a raw SUBREG if simplify_gen_subreg fails.

gcc/testsuite/ChangeLog:

	* gcc.dg/pr123295-1.c: New test.

Signed-off-by: Andrew Pinski <andrew.pinski@oss.qualcomm.com>