doc, rl78: Clean up RL78 option documentation [PR122243] [PR71340]
This target is probably dead (no listed maintainer and no support for
LRA).  There seems to be no interest in reviving the 2014 patch for
the documented but never-committed -m64bit-doubles and -m32bit-doubles
options, so I've removed those docs instead, along with some other
routine housekeeping to keep the docs in sync with the options file.

gcc/ChangeLog
	PR other/122243
	PR target/71340
	* doc/invoke.texi (Option Summary) <RL78 Options>: Remove
	never-implemented -m64bit-doubles and -m32bit-doubles options.
	Add missing entries for -mrelax, -mes0, -mmul=rl78, -mcpu=rl78,
	and -mrl78.
	(RL78 Options): Likewise.  Add missing @opindex entries.
	Light copy-editing.