testsuite: Skip pr123295-1.c for non int128 targets [PR123334]
This was an oversight on my part. The testcase uses __int128 but
forgot to check if it is compiling for a target that supports that
type.

Mark the testcase as unsupported for non-int128 targets.
Pushed as obvious after run the testcase with and without -m32 on x86_64:
make check-gcc RUNTESTFLAGS="--target_board=unix/-m32 dg.exp=pr123295-1.c"
make check-gcc RUNTESTFLAGS="--target_board=unix dg.exp=pr123295-1.c"

	PR testsuite/123334
gcc/testsuite/ChangeLog:

	* gcc.dg/pr123295-1.c: Require int128.

Signed-off-by: Andrew Pinski <andrew.pinski@oss.qualcomm.com>